New textures for trees and rocks + a little bit of grass and incresed the size of the crow nest to 1.5 of the previous size
Added a temporary folder for the trees and placed them in the scene you can see what they look like (turned out a bit too desaturated and yellow/red perhaps). Fixed the rocks UV's and added the new textures :) Added a bit of grass around the crow nest tree to see what it looks like. Deleted old rock textures. Updated the crow nest mesh in Maya, made it 1.5 in scale and fixed the UVs a little.  It should now be easier to hit (might require fixes to the coliders and scrips)