NEW FRICKING GRASS! New Fountain
Fixed new grass, looks way better but still need to adjust the colors a bit. Fixed a new fountain coz the old one was just a pice of shit anyway. Don't fool yourself, the new one is also a pile of shit but it has water and easier UVs to fix